! A tiny int breaks big things (Postgresql/Sqlite sql files) [Bug 3672]
git-svn-id: https://devel.simplemachines.org/code/smf/trunk@8904 a96bbafc-d752-43aa-a6db-eed64a5f2cf9

+++ other/install_2-0_sqlite.sql
@@ -1258,7 +1258,7 @@ CREATE TABLE {$db_prefix}mail_queue (
   headers text NOT NULL,
   send_html smallint NOT NULL default '0',
   priority smallint NOT NULL default '1',
-  private tinyint(1) NOT NULL default '0'
+  private smallint NOT NULL default '0'
 );
 
 #
